Combat Movement Utility AI & Triggers
Combat Movement Utility AI prototype has been completed and is working. Decoupling of Utility AI and particular systems is desireable. Fixed bugs in trigger system leading to too many trigger events raised.

+++ Assets/Game Scripts/Main/AI/Consideration.cs
@@ -8,8 +8,8 @@ public struct Consideration : IComponentData
     public GraphType GraphType;
     public float Slope;
     public float Exp;
-    public float XShift;
-    public float YShift;
+    public half XShift;
+    public half YShift;
 
     public float Evaluate(float input)
     {
